[sarlc] Code refactoring.
